Control added in miRNA_detection

+++ daemon.sh
@@ -11,12 +11,15 @@ fi
 export PATH=$CODE_PATH'/aux_sh:'$PATH
 export PATH=$CODE_PATH'/aux_parsers:'$PATH
 source $CONFIG_DAEMON
-rm -r $TARGETS_FOLDER
-mkdir $TARGETS_FOLDER
-eval "$generate_targets"
-export TARGETS=`ls $TARGETS_FOLDER/*_target.txt | rev | cut -f 1 -d "/" | rev | tr "\n" ","` ; TARGETS=${TARGETS%?}	#-------#	Target file location, including a short sample description	
-n_target=`echo $TARGETS |tr "," "\n" | wc -l `
-tasks=`echo $n_target"+1" | bc`
+if [ $experiment_type != "miRNAseq_detection" ] ; then 
+
+	rm -r $TARGETS_FOLDER
+	mkdir $TARGETS_FOLDER
+	eval "$generate_targets"
+	export TARGETS=`ls $TARGETS_FOLDER/*_target.txt | rev | cut -f 1 -d "/" | rev | tr "\n" ","` ; TARGETS=${TARGETS%?}	#-------#	Target file location, including a short sample description	
+	n_target=`echo $TARGETS |tr "," "\n" | wc -l `
+	tasks=`echo $n_target"+1" | bc`
+fi
 
 ## STAGE EXECUTION
 #######################################################################
